+++ qiskit/backends/aer/__init__.py
@@ -13,7 +13,7 @@ from .qasm_simulator import CliffordSimulator, QasmSimulator
 from .qasm_simulator_py import QasmSimulatorPy
 from .statevector_simulator import StatevectorSimulator
 from .statevector_simulator_py import StatevectorSimulatorPy
-from .unitary_simulator import UnitarySimulator
+from .unitary_simulator_py import UnitarySimulatorPy
 
 # Global instance to be used as the entry point for convenience.
 Aer = AerProvider()  # pylint: disable=invalid-name