+++ CHANGELOG.rst
@@ -66,6 +66,7 @@ Changed
 - Backend and Provider methods defined in the specification use model objects
   rather than dicts, along with validation against schemas (#1249, #1277).
 - ``backend.provider()`` is now a method instead of a property (#1312).
+- Remove local backend (Aer) fallback (#1303)
 
 Deprecated
 """"""""""
@@ -83,6 +84,8 @@ Deprecated
 - The ``plot_barriers`` and ``reverse_bits`` keys in the ``style`` kwarg dict
   are deprecated, instead the `qiskit.tools.visualization.circuit_drawer()`
   kwargs ``plot_barriers`` and ``reverse_bits`` should be used instead. (#1180)
+- The transpiler methods do not support emitting multiple output `format`
+  anymore (#1319).
 
 
 Removed